+++ lib/theme_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_job/ui/res/app_colors.dart';
-import 'package:flutter_job/ui/screen/res/themes.dart';
+import 'package:flutter_job/ui/screens/res/themes.dart';
 
 class ThemeProvider extends ChangeNotifier {
   set isLightTheme(bool value) {
