+++ lib/ui/screens/new_place_category.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_job/constants.dart';
 import 'package:flutter_job/main.dart';
 import 'package:flutter_job/ui/res/app_assets.dart';
 import 'package:flutter_job/ui/res/app_strings.dart';
 import 'package:flutter_job/ui/res/app_typography.dart';
-import 'package:flutter_job/ui/screen/add_sight_screen.dart';
 import 'package:flutter_svg/svg.dart';
+
 
 AppTypography appTypography = AppTypography();
