+++ lib/ui/screens/sight_card_visited.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_job/domain/sight.dart';
 import 'package:flutter_job/main.dart';
@@ -7,18 +6,18 @@ import 'package:flutter_job/ui/res/app_assets.dart';
 import 'package:flutter_job/ui/res/app_typography.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class SightCardPlan extends StatefulWidget {
+class SightCardVisited extends StatefulWidget {
   final Sight? sight;
 
-  const SightCardPlan({Key? key, required this.sight}) : super(key: key);
+  const SightCardVisited({Key? key, required this.sight}) : super(key: key);
 
   @override
-  State<SightCardPlan> createState() => _SightCardPlanState();
+  State<SightCardVisited> createState() => _SightCardVisitedState();
 }
 
 AppTypography appTypography = AppTypography();
 
-class _SightCardPlanState extends State<SightCardPlan> {
+class _SightCardVisitedState extends State<SightCardVisited> {
   @override
   Widget build(BuildContext context) {
     return Padding(
@@ -69,14 +68,10 @@ class _SightCardPlanState extends State<SightCardPlan> {
                         CupertinoButton(
                           padding: const EdgeInsets.all(10),
                           child: SvgPicture.asset(
-                            AppAssets.calendar,
+                            AppAssets.share,
                             color: themeProvider.appTheme.whiteColor,
                           ),
-                          onPressed: () {
-                            if (kDebugMode) {
-                              print('Calendar Pressed');
-                            }
-                          },
+                          onPressed: () {},
                         ),
                         CupertinoButton(
                           padding: const EdgeInsets.all(10),
@@ -84,11 +79,7 @@ class _SightCardPlanState extends State<SightCardPlan> {
                             AppAssets.close,
                             color: themeProvider.appTheme.whiteColor,
                           ),
-                          onPressed: () {
-                            if (kDebugMode) {
-                              print('Close Pressed');
-                            }
-                          },
+                          onPressed: () {},
                         ),
                       ],
                     ),
@@ -113,7 +104,6 @@ class _SightCardPlanState extends State<SightCardPlan> {
                   ),
                 ),
                 height: 100,
-                width: double.infinity,
                 child: Column(
                   children: [
                     Container(
@@ -137,9 +127,8 @@ class _SightCardPlanState extends State<SightCardPlan> {
                       ),
                       alignment: Alignment.topLeft,
                       child: Text(
-                        'Запланировано на 12 окт. 2022',
-                        style: appTypography.text14Regular
-                            .copyWith(color: themeProvider.appTheme.greenColor),
+                        'Цель достигнута 12 окт. 2022',
+                        style: appTypography.textGreyInactive14Regular,
                       ),
                     ),
                   ],
